safari-534.53-branch/LayoutTests/inspector/console/console-clear.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function log() {
    // Fill console.
    console.log("one");
    console.log("two");
    console.log("three");
}

log();

function test()
{

    InspectorTest.addResult("=== Before clear ==="); 
    InspectorTest.dumpConsoleMessages();

    WebInspector.console.requestClearMessages();
    
    function callback()
    {
        InspectorTest.addResult("=== After clear ==="); 
        InspectorTest.dumpConsoleMessages();
        InspectorTest.completeTest();
    }
    InspectorTest.runAfterPendingDispatches(callback);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that console is cleared upon requestClearMessages call.
</p>

</body>
</html>
