LayoutTests/inspector/console/console-clear-expected.txt

CONSOLE MESSAGE: line 9: one
CONSOLE MESSAGE: line 10: two
CONSOLE MESSAGE: line 11: three
Tests that console is cleared upon requestClearMessages call.

=== Before clear ===
console-clear.html:9one
console-clear.html:10two
console-clear.html:11three
=== After clear ===
